% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{addColumns}
\alias{addColumns}
\title{Add Columns to RMarkdown-Presentations}
\usage{
addColumns(width = 50)
}
\arguments{
\item{width}{width of first column}
}
\description{
Add Columns to RMarkdown-Presentations
}
