% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{ConvertWindowsPath}
\alias{ConvertWindowsPath}
\title{Converts single backslashes to forward slashes (e.g. in Windows paths)}
\usage{
ConvertWindowsPath()
}
\description{
Converts single backslashes to forward slashes (e.g. in Windows paths)
}
