% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{extend_range}
\alias{extend_range}
\title{extend_range}
\usage{
extend_range(doc_range, ext)
}
\arguments{
\item{doc_range}{(single) document_range object}

\item{ext}{number of lines to be extend, positive downward}
}
\value{
a new list of extended range(s)
}
\description{
Extend the current single-line range to lines above or below,
    spanning the same columns
}

