% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{extend_selection}
\alias{extend_selection}
\title{extend_selection}
\usage{
extend_selection(ext)
}
\arguments{
\item{ext}{number of lines to be extend, positive means downward}
}
\value{
new extened ranges (including the old ones)
}
\description{
extend current selection by `ext` number of lines
}
