% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{from_import}
\alias{from_import}
\title{from_import imports specific objects from a package}
\usage{
from_import(pkg, objs, as = NULL, .hidden = F)
}
\arguments{
\item{pkg}{package name, a string}

\item{objs}{a vector of obj names}

\item{as}{aliases of the object in the current environment}

\item{.hidden}{logical, import the hidden obj or not}
}
\description{
It attempts to mimic python's `import ... from ...` construct
}
\examples{
from_import('dplyr', 'select', 'dselect')
from_import('AnnotationDbi', 'select', 'aselect')
}
