% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspection.R
\name{get_name_near_cursor}
\alias{get_name_near_cursor}
\title{Get the object name near the cursor.}
\usage{
get_name_near_cursor()
}
\value{
The name of the object. \code{NULL} if none is under cursor.
}
\description{
Get the object name near the cursor.
}
\seealso{
\code{\link{echo_class}}
}
\keyword{internal}
