% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertCompundPipeAddin.R
\name{insertCompoundPipeAddin}
\alias{insertCompoundPipeAddin}
\title{Insert \%<>\%.}
\usage{
insertCompoundPipeAddin()
}
\description{
Call this function as an addin to insert \code{ \%in\% } at the cursor position.
}
