% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertLibraryCallsAddin.R
\name{insertLibraryCallsAddin}
\alias{insertLibraryCallsAddin}
\title{Insert some often used libraries (dplyr, magri)}
\usage{
insertLibraryCallsAddin()
}
\description{
Call this function as an addin to insert \code{library(...);library(...);...} at the cursor position.
}
