% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{remove_range}
\alias{remove_range}
\title{remove_range}
\usage{
remove_range(doc_ranges, ext)
}
\arguments{
\item{doc_ranges}{list of document_ranges objects}

\item{ext}{number of line}
}
\value{
new ranges
}
\description{
remove ranges by ext
}

