% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{remove_selection}
\alias{remove_selection}
\title{remove_selection}
\usage{
remove_selection(ext)
}
\arguments{
\item{ext}{number of lines to be removed, positive means from bottom}
}
\value{
newly shrunken ranges
}
\description{
remove current selection by `ext` amount
}
