% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{replaceSelectedText}
\alias{replaceSelectedText}
\title{Replaces selected text in actual source editor}
\usage{
replaceSelectedText(search, replace)
}
\arguments{
\item{search}{text to be replaced}

\item{replace}{replacement text}
}
\description{
Replaces selected text in actual source editor
}
