% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio_section.R
\name{rstudio_section}
\alias{rstudio_section}
\title{Converts text to an RStudio section}
\usage{
rstudio_section()
}
\description{
This addon converts a text to an RStudio section or code folding block.
}

