% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{ConvertWindowsPath}
\alias{ConvertWindowsPath}
\title{Converts single backslashes from windows path to forward slashes}
\usage{
ConvertWindowsPath()
}
\description{
Converts single backslashes from windows path to forward slashes
}
